+++ lib/storage/search.go
@@ -187,8 +187,11 @@ func (s *Search) reset() {
 //
 // Init returns the upper bound on the number of found time series.
 func (s *Search) Init(qt *querytracer.Tracer, storage *Storage, tfss []*TagFilters, tr TimeRange, maxMetrics int, deadline uint64) int {
-	qt = qt.NewChild("init series search: filters=%s, timeRange=%s", tfss, &tr)
-	defer qt.Done()
+	sc, so, cleanup := getCommonSearchOptions(qt, deadline, maxMetrics, "search")
+	defer cleanup()
+
+	sc = sc.NewChild("init series search: filters=%s, timeRange=%s", tfss, &tr)
+	defer sc.Done()
 
 	indexTR := storage.adjustTimeRange(tr)
 	dataTR := tr
@@ -203,24 +206,24 @@ func (s *Search) Init(qt *querytracer.Tracer, storage *Storage, tfss []*TagFilte
 	s.retentionDeadline = retentionDeadline
 	s.tr = tr
 	s.tfss = tfss
-	s.deadline = deadline
+	s.deadline = sc.deadline
 	s.needClosing = true
 
 	var tsids []TSID
-	metricIDs, err := s.idb.searchMetricIDs(qt, tfss, indexTR, maxMetrics, deadline)
+	metricIDs, err := s.idb.searchMetricIDs(sc, tfss, indexTR, so)
 	if err == nil && len(metricIDs) > 0 && len(tfss) > 0 {
 		accountID := tfss[0].accountID
 		projectID := tfss[0].projectID
-		tsids, err = s.idb.getTSIDsFromMetricIDs(qt, accountID, projectID, metricIDs, deadline)
+		tsids, err = s.idb.getTSIDsFromMetricIDs(sc, accountID, projectID, metricIDs)
 		if err == nil {
-			err = storage.prefetchMetricNames(qt, s.idb, accountID, projectID, metricIDs, deadline)
+			err = storage.prefetchMetricNames(sc, s.idb, accountID, projectID, metricIDs)
 		}
 	}
 	// It is ok to call Init on non-nil err.
 	// Init must be called before returning because it will fail
 	// on Search.MustClose otherwise.
 	s.ts.Init(storage.tb, tsids, dataTR)
-	qt.Printf("search for parts with data for %d series", len(tsids))
+	sc.Printf("search for parts with data for %d series", len(tsids))
 	if err != nil {
 		s.err = err
 		return 0
@@ -596,3 +599,13 @@ const (
 	paceLimiterMediumIterationsMask = 1<<14 - 1
 	paceLimiterSlowIterationsMask   = 1<<12 - 1
 )
+
+func getCommonSearchOptions(qt *querytracer.Tracer, deadline uint64, maxMetrics int, source string) (*searchContext, *searchOptions, func()) {
+	sc := getSearchContext(deadline, source, qt)
+	so := getSearchOptions(maxMetrics)
+
+	return sc, so, func() {
+		putSearchContext(sc)
+	}
+
+}